+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.132.0] - Not released
+### Fixed
+- Fixed a draft fileIds synchronization bug.
 
 ## [1.131.0] - 2024-05-05
 ### Added